+++ lib/built_in_plugins/device_info_plugin.dart
@@ -1,81 +1,49 @@
-import 'dart:io' show Platform;
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
+import 'package:flutter_dev_toolkit/built_in_plugins/widgets/device_info_tab.dart';
+import 'package:flutter_dev_toolkit/core/device_info_store.dart';
 import 'package:share_plus/share_plus.dart';
 
 import '../core/dev_toolkit_plugin.dart';
 import '../flutter_dev_toolkit.dart';
 
-class DeviceInfoPlugin extends DevToolkitPlugin {
+class UserDeviceInfoPlugin extends DevToolkitPlugin {
   @override
   String get name => 'Device Info';
 
   @override
   IconData get icon => Icons.devices;
-  Map<String, dynamic> info = {};
 
   @override
   void onInit() {}
 
   @override
   List<Widget> buildActions(BuildContext context) {
-    ElevatedButton.icon(
-      icon: const Icon(Icons.upload_file),
-      label: const Text('Export'),
-      onPressed: () {
-        String shareData = '';
+    return [
+      IconButton(
+        onPressed: () {
+          Map<String, dynamic> shareData = {};
 
-        for (var element in info.entries) {
-          shareData += '${element.key} : ${element.value}\n';
-        }
-        FlutterDevToolkit.logger.log('📤 Exported Device Info:\n$shareData');
-        ShareParams shareParams = ShareParams(text: shareData, subject: name);
-        SharePlus.instance.share(shareParams);
-      },
-    );
-
-    return super.buildActions(context);
+          for (var element in DeviceInfoLogStore.logs) {
+            // shareData += '${element.key} : ${element.value}\n';
+            // shareData.add(element.toJson());
+            shareData[element.key] = element.value;
+          }
+          FlutterDevToolkit.logger.log('📤 Exported Device Info:\n$shareData');
+          ShareParams shareParams = ShareParams(
+            text: jsonEncode(shareData),
+            subject: name,
+          );
+          SharePlus.instance.share(shareParams);
+        },
+        icon: Icon(Icons.share),
+      ),
+    ];
   }
 
   @override
   Widget buildTab(BuildContext context) {
-    final mq = MediaQuery.of(context);
-
-    info = {
-      'Platform': Platform.operatingSystem,
-      'OS Version': Platform.operatingSystemVersion,
-      'Screen Size': '${mq.size.width} x ${mq.size.height}',
-      'Pixel Ratio': '${mq.devicePixelRatio}',
-      'Orientation': mq.orientation.name,
-      'Locale': Localizations.localeOf(context).toLanguageTag(),
-    };
-
-    return Column(
-      children: [
-        Expanded(
-          child: ListView(
-            padding: const EdgeInsets.all(16),
-            children:
-                info.entries.map((e) {
-                  return Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        e.key,
-                        style: const TextStyle(fontWeight: FontWeight.bold),
-                      ),
-                      Text(e.value),
-                      const Divider(),
-                    ],
-                  );
-                }).toList(),
-          ),
-        ),
-
-        Row(
-          mainAxisAlignment: MainAxisAlignment.end,
-          children: [const SizedBox(width: 12)],
-        ),
-      ],
-    );
+    return DeviceInfoTab();
   }
 }